Simplify SchemaField, by removing bloat.
This means remove concept of materialisation, SchemaField is no longer
lazy-loading so field is not concerned when and if it is materialised.

Another feature that is being dropped is inline declaration of
SchemaField. SchemaField is now exclusively Declarative, you define a
class with additional configuration (like throw exception on unknown
key) but there is no way to define SchemaField object on the fly by
passing all the arguments to constructor. It is important as it was a
hurdle when combining declarative classes with inline classes and still
working out to resolve inheritance and overriding parameters on MRO
resolution.

As a general rule of thumb constructor sets now only attributes that
influence given class characteristic on runtime (think validation) and
not defines what the class does (like passing list of fields that given
schema accepts).

Also I am redesigining tests and then code, working TDD way, so that I
will end up with something more intuitive and more real-life than ever.

This is the first commit in series.