Create new field type - lazy field. It will handle cycles.
Previously SchemaField could be used in two ways as-is or to lazy-load
another SchemaField it added complexity and serve little purpose, there
was one actually that it would make introspeciot easier as field would
be lazy or not but always SchemaField with certain structure, so it's
gone. This is initial commit as much of lazyfield will be developed in
following commits and it may end up not as a Field but lazy_loader as a
function.