Stop overriding name on instances of field_type.
Previously there could be possibility that field_type was not an
instance and as such could not have name defined. But from now on it's
always an instance and name is checked by constractor and required to be
set to some value, so only part missing on name is index of the
element on a collection.